requirements.txt
asgiref==3.10.0
Django==5.2.7
mysqlclient==2.2.7
pymysql==1.1.2
psycopg2-binary==2.9.9
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
python-decouple==3.8
python-dotenv==1.1.1
requests==2.31.0
google-generativeai==0.3.2
